Refactoring
move interceptors from macaroon service
move permissions from main
reorg macaroons files inside data dir
clean up main and test
