+++ cmd/grpc/main.go
@@ -1,166 +1,29 @@
 package main
 
 import (
-	"context"
-	"encoding/hex"
 	"fmt"
+	"github.com/improbable-eng/grpc-web/go/grpcweb"
+	log "github.com/sirupsen/logrus"
+	"github.com/soheilhy/cmux"
+	"github.com/tdex-network/tdex-daemon/config"
+	"github.com/tdex-network/tdex-daemon/internal/core/application"
 	dbbadger "github.com/tdex-network/tdex-daemon/internal/infrastructure/storage/db/badger"
+	grpchandler "github.com/tdex-network/tdex-daemon/internal/interfaces/grpc/handler"
+	"github.com/tdex-network/tdex-daemon/internal/interfaces/grpc/interceptor"
+	"github.com/tdex-network/tdex-daemon/pkg/crawler"
+	"github.com/tdex-network/tdex-daemon/pkg/explorer"
 	"github.com/tdex-network/tdex-daemon/pkg/macaroons"
-	"gopkg.in/macaroon-bakery.v2/bakery"
-	"io/ioutil"
+	pbhandshake "github.com/tdex-network/tdex-protobuf/generated/go/handshake"
+	pboperator "github.com/tdex-network/tdex-protobuf/generated/go/operator"
+	pbtrader "github.com/tdex-network/tdex-protobuf/generated/go/trade"
+	pbwallet "github.com/tdex-network/tdex-protobuf/generated/go/wallet"
+	"google.golang.org/grpc"
 	"net"
 	"net/http"
 	"os"
 	"os/signal"
 	"path/filepath"
 	"syscall"
-
-	dbbadger "github.com/tdex-network/tdex-daemon/internal/infrastructure/storage/db/badger"
-
-	"github.com/improbable-eng/grpc-web/go/grpcweb"
-	"github.com/soheilhy/cmux"
-	"github.com/tdex-network/tdex-daemon/internal/core/application"
-	grpchandler "github.com/tdex-network/tdex-daemon/internal/interfaces/grpc/handler"
-	"github.com/tdex-network/tdex-daemon/internal/interfaces/grpc/interceptor"
-
-	log "github.com/sirupsen/logrus"
-	"github.com/tdex-network/tdex-daemon/config"
-	"github.com/tdex-network/tdex-daemon/pkg/crawler"
-	"github.com/tdex-network/tdex-daemon/pkg/explorer"
-	"google.golang.org/grpc"
-
-	pbhandshake "github.com/tdex-network/tdex-protobuf/generated/go/handshake"
-	pboperator "github.com/tdex-network/tdex-protobuf/generated/go/operator"
-	pbtrader "github.com/tdex-network/tdex-protobuf/generated/go/trade"
-	pbwallet "github.com/tdex-network/tdex-protobuf/generated/go/wallet"
-)
-
-var (
-	pricePermissions = []bakery.Op{
-		{
-			Entity: "market",
-			Action: "updatemarketprice",
-		},
-	}
-	marketPermissions = []bakery.Op{
-		{
-			Entity: "market",
-			Action: "openmarket",
-		},
-		{
-			Entity: "market",
-			Action: "closemarket",
-		},
-		{
-			Entity: "market",
-			Action: "updatemarketstrategy",
-		},
-	}
-	//TODO check permissions
-	readonlyPermissions = []bakery.Op{}
-
-	adminPermissions = []bakery.Op{
-		{
-			Entity: "wallet",
-			Action: "genseed",
-		},
-		{
-			Entity: "wallet",
-			Action: "initwallet",
-		},
-		{
-			Entity: "wallet",
-			Action: "unlockwallet",
-		},
-		{
-			Entity: "wallet",
-			Action: "changepassword",
-		},
-		{
-			Entity: "wallet",
-			Action: "walletaddress",
-		},
-		{
-			Entity: "wallet",
-			Action: "walletbalance",
-		},
-		{
-			Entity: "wallet",
-			Action: "sendtomany",
-		},
-		{
-			Entity: "operator",
-			Action: "depositmarket",
-		},
-		{
-			Entity: "operator",
-			Action: "listdepositmarket",
-		},
-		{
-			Entity: "operator",
-			Action: "depositfeeaccount",
-		},
-		{
-			Entity: "operator",
-			Action: "balancefeeaccount",
-		},
-		{
-			Entity: "operator",
-			Action: "openmarket",
-		},
-		{
-			Entity: "operator",
-			Action: "closemarket",
-		},
-		{
-			Entity: "operator",
-			Action: "listmarket",
-		},
-		{
-			Entity: "operator",
-			Action: "updatemarketfee",
-		},
-		{
-			Entity: "operator",
-			Action: "updatemarketprice",
-		},
-		{
-			Entity: "operator",
-			Action: "updatemarketstrategy",
-		},
-		{
-			Entity: "operator",
-			Action: "withdrawmarket",
-		},
-		{
-			Entity: "operator",
-			Action: "listswaps",
-		},
-		{
-			Entity: "operator",
-			Action: "reportmarketfee",
-		},
-		{
-			Entity: "trade",
-			Action: "markets",
-		},
-		{
-			Entity: "trade",
-			Action: "balances",
-		},
-		{
-			Entity: "trade",
-			Action: "marketprice",
-		},
-		{
-			Entity: "trade",
-			Action: "tradepropose",
-		},
-		{
-			Entity: "trade",
-			Action: "tradecomplete",
-		},
-	}
 )
 
 func main() {
@@ -172,10 +35,9 @@ func main() {
 		log.WithError(err).Panic("error while opening db")
 	}
 
-	// ****
 	var macaroonService *macaroons.Service
 	macaroonService, err = macaroons.NewService(
-		config.GetString(config.DataDirPathKey),
+		filepath.Join(config.GetString(config.DataDirPathKey), "macaroon"),
 		"tdex-daemon",
 		macaroons.IPLockChecker,
 	)
@@ -184,46 +46,17 @@ func main() {
 	}
 	defer macaroonService.Close()
 
-	var pass = []byte("hello")
+	var pass = []byte("hello") //TODO this should be passed as flag/env?
 	// Try to unlock the macaroon store with the private password.
 	err = macaroonService.CreateUnlock(&pass)
 	if err != nil {
 		log.WithError(err).Panic("unable to unlock macaroons")
 	}
 
-	adminMacPath := filepath.Join(
-		config.GetString(config.DataDirPathKey), "admin.macaroon",
-	)
-
-	readMacPath := filepath.Join(
-		config.GetString(config.DataDirPathKey), "readonly.macaroon",
-	)
-
-	priceMacPath := filepath.Join(
-		config.GetString(config.DataDirPathKey), "price.macaroon",
-	)
-
-	marketMacPath := filepath.Join(
-		config.GetString(config.DataDirPathKey), "market.macaroon",
-	)
-
-	// Create macaroon files for lncli to use if they don't exist.
-	if !fileExists(adminMacPath) && !fileExists(readMacPath) &&
-		!fileExists(priceMacPath) && !fileExists(marketMacPath) {
-
-		err = genMacaroons(
-			context.Background(),
-			macaroonService,
-			adminMacPath,
-			readMacPath,
-			priceMacPath,
-			marketMacPath,
-		)
-		if err != nil {
-			log.WithError(err).Panic("unable to create macaroons")
-		}
+	err = macaroonService.Init()
+	if err != nil {
+		log.WithError(err).Panic("unable to init macaroon")
 	}
-	// ****
 
 	unspentRepository := dbbadger.NewUnspentRepositoryImpl(dbManager)
 	vaultRepository := dbbadger.NewVaultRepositoryImpl(dbManager)
@@ -271,11 +104,11 @@ func main() {
 	operatorAddress := fmt.Sprintf(":%+v", config.GetInt(config.OperatorListeningPortKey))
 	// Grpc Server
 	traderGrpcServer := grpc.NewServer(
-		interceptor.UnaryInterceptor(dbManager, macaroonService),
+		interceptor.UnaryInterceptor(dbManager, *macaroonService),
 		interceptor.StreamInterceptor(dbManager, macaroonService),
 	)
 	operatorGrpcServer := grpc.NewServer(
-		interceptor.UnaryInterceptor(dbManager, macaroonService),
+		interceptor.UnaryInterceptor(dbManager, *macaroonService),
 		interceptor.StreamInterceptor(dbManager, macaroonService),
 	)
 
@@ -355,93 +188,3 @@ func serveMux(address string, grpcServer *grpc.Server) error {
 	go mux.Serve()
 	return nil
 }
-
-// ****
-
-// genMacaroons generates three macaroon files; one admin-level, one for
-// invoice access and one read-only. These can also be used to generate more
-// granular macaroons.
-func genMacaroons(ctx context.Context, svc *macaroons.Service,
-	admFile, roFile, priceFile, marketFile string) error {
-
-	priceMac, err := svc.NewMacaroon(
-		ctx, macaroons.DefaultRootKeyID, pricePermissions...,
-	)
-	if err != nil {
-		return err
-	}
-	priceUpdateMacBytes, err := priceMac.M().MarshalBinary()
-	if err != nil {
-		return err
-	}
-	err = ioutil.WriteFile(priceFile, priceUpdateMacBytes, 0644)
-	if err != nil {
-		os.Remove(priceFile)
-		return err
-	}
-
-	marketMac, err := svc.NewMacaroon(
-		ctx, macaroons.DefaultRootKeyID, marketPermissions...,
-	)
-	if err != nil {
-		return err
-	}
-	marketMacBytes, err := marketMac.M().MarshalBinary()
-	if err != nil {
-		return err
-	}
-	err = ioutil.WriteFile(marketFile, marketMacBytes, 0644)
-	if err != nil {
-		os.Remove(marketFile)
-		return err
-	}
-
-	//TODO uncomment once read only permissions are defined
-	//// Generate the read-only macaroon and write it to a file.
-	//roMacaroon, err := svc.NewMacaroon(
-	//	ctx, macaroons.DefaultRootKeyID, readonlyPermissions...,
-	//)
-	//if err != nil {
-	//	return err
-	//}
-	//roBytes, err := roMacaroon.M().MarshalBinary()
-	//if err != nil {
-	//	return err
-	//}
-	//if err = ioutil.WriteFile(roFile, roBytes, 0644); err != nil {
-	//	os.Remove(admFile)
-	//	return err
-	//}
-
-	// Generate the admin macaroon and write it to a file.
-	admMacaroon, err := svc.NewMacaroon(
-		ctx, macaroons.DefaultRootKeyID, adminPermissions...,
-	)
-	if err != nil {
-		return err
-	}
-	admBytes, err := admMacaroon.M().MarshalBinary()
-	if err != nil {
-		return err
-	}
-	println(hex.EncodeToString(admBytes))
-
-	if err = ioutil.WriteFile(admFile, admBytes, 0600); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// fileExists reports whether the named file or directory exists.
-// This function is taken from https://github.com/btcsuite/btcd
-func fileExists(name string) bool {
-	if _, err := os.Stat(name); err != nil {
-		if os.IsNotExist(err) {
-			return false
-		}
-	}
-	return true
-}
-
-// ****
