+++ internal/interfaces/grpc/interceptor/interceptor.go
@@ -5,17 +5,16 @@ import (
 	dbbadger "github.com/tdex-network/tdex-daemon/internal/infrastructure/storage/db/badger"
 	"github.com/tdex-network/tdex-daemon/pkg/macaroons"
 	"google.golang.org/grpc"
-	"gopkg.in/macaroon-bakery.v2/bakery"
 )
 
 // UnaryInterceptor returns the unary interceptor
 func UnaryInterceptor(
 	dbManager *dbbadger.DbManager,
-	macaroonService *macaroons.Service,
+	macaroonService macaroons.Service,
 ) grpc.ServerOption {
 	return grpc.UnaryInterceptor(
 		middleware.ChainUnaryServer(
-			macaroonService.UnaryServerInterceptor(MainRPCServerPermissions()),
+			unaryAuthHandler(macaroons.RPCServerPermissions(), macaroonService),
 			unaryLogger,
 			unaryTransactionHandler(dbManager),
 		),
@@ -29,116 +28,9 @@ func StreamInterceptor(
 ) grpc.ServerOption {
 	return grpc.StreamInterceptor(
 		middleware.ChainStreamServer(
-			macaroonService.StreamServerInterceptor(MainRPCServerPermissions()),
+			streamAuthHandler(macaroons.RPCServerPermissions(), *macaroonService),
 			streamLogger,
 			streamTransactionHandler(dbManager),
 		),
 	)
-}
-
-// MainRPCServerPermissions returns a mapping of the main RPC server calls to
-// the permissions they require.
-func MainRPCServerPermissions() map[string][]bakery.Op {
-	return map[string][]bakery.Op{
-		"/Wallet/GenSeed": {{
-			Entity: "wallet",
-			Action: "genseed",
-		}},
-		"/Wallet/InitWallet": {{
-			Entity: "wallet",
-			Action: "initwallet",
-		}},
-		"/Wallet/UnlockWallet": {{
-			Entity: "wallet",
-			Action: "unlockwallet",
-		}},
-		"/Wallet/ChangePassword": {{
-			Entity: "wallet",
-			Action: "changepassword",
-		}},
-		"/Wallet/WalletAddress": {{
-			Entity: "wallet",
-			Action: "walletaddress",
-		}},
-		"/Wallet/WalletBalance": {{
-			Entity: "wallet",
-			Action: "walletbalance",
-		}},
-		"/Wallet/SendToMany": {{
-			Entity: "wallet",
-			Action: "sendtomany",
-		}},
-		"/Operator/DepositMarket": {{
-			Entity: "operator",
-			Action: "depositmarket",
-		}},
-		"/Operator/ListDepositMarket": {{
-			Entity: "operator",
-			Action: "listdepositmarket",
-		}},
-		"/Operator/DepositFeeAccount": {{
-			Entity: "operator",
-			Action: "depositfeeaccount",
-		}},
-		"/Operator/BalanceFeeAccount": {{
-			Entity: "operator",
-			Action: "balancefeeaccount",
-		}},
-		"/Operator/OpenMarket": {{
-			Entity: "operator",
-			Action: "openmarket",
-		}},
-		"/Operator/CloseMarket": {{
-			Entity: "operator",
-			Action: "closemarket",
-		}},
-		"/Operator/ListMarket": {{
-			Entity: "operator",
-			Action: "listmarket",
-		}},
-		"/Operator/UpdateMarketFee": {{
-			Entity: "operator",
-			Action: "updatemarketfee",
-		}},
-		"/Operator/UpdateMarketPrice": {{
-			Entity: "operator",
-			Action: "updatemarketprice",
-		}},
-		"/Operator/UpdateMarketStrategy": {{
-			Entity: "operator",
-			Action: "updatemarketstrategy",
-		}},
-		"/Operator/WithdrawMarket": {{
-			Entity: "operator",
-			Action: "withdrawmarket",
-		}},
-		"/Operator/ListSwaps": {{
-			Entity: "operator",
-			Action: "listswaps",
-		}},
-		"/Operator/ReportMarketFee": {{
-			Entity: "operator",
-			Action: "reportmarketfee",
-		}},
-		"/Trade/Markets": {{
-			Entity: "trade",
-			Action: "markets",
-		}},
-		"/Trade/Balances": {{
-			Entity: "trade",
-			Action: "balances",
-		}},
-		"/Trade/MarketPrice": {{
-			Entity: "trade",
-			Action: "marketprice",
-		}},
-		"/Trade/TradePropose": {{
-			Entity: "trade",
-			Action: "tradepropose",
-		}},
-		"/Trade/TradeComplete": {{
-			Entity: "trade",
-			Action: "tradecomplete",
-		}},
-	}
 }
